проблемы с паролем и авторизаций

+++ src/main/java/ru/kata/spring/boot_security/demo/repository/UserRepository.java
@@ -1,12 +1,14 @@
 package ru.kata.spring.boot_security.demo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import ru.kata.spring.boot_security.demo.model.User;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
+    //@Query("SELECT u FROM User u JOIN FETCH u.roles WHERE u.username = :username")
     Optional<User> findByUsername(String username);
 }
